pyproject.toml
[project]
name = "dict2md"
version = "0.1.0"
description = "A function that takes a dictionary and renders it as a markdown string."
authors = [{ name = "Chris", email = "chris@cjrobson.com" }]
dependencies = [
    "python-docx>=1.1.2",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/dict2md"]
